+++ server/migrations/0010_add_fields_to_market_location.up.sql
@@ -1,5 +1,5 @@
 START TRANSACTION;
 
-ALTER TABLE fp_market_locations add COLUMN esi_character_id int;
+ALTER TABLE fp_market_locations add COLUMN esi_character_id bigint;
 
 commit;
